+++ src/components/blank-page/Blank.tsx
@@ -1,9 +1,16 @@
+import { Link } from 'react-router-dom';
+import { formPath } from '../form/Form';
 import './Blank.css';
 
+export const blankPagePath = 'blank'; 
+
 function BlankPage(): JSX.Element {
-    return (
-        <div className='Header'>Teste</div>
-    );
+  return (
+    <>
+      <h1 className='Header'>Teste</h1>
+      <Link to={formPath}>Form</Link>
+    </>
+  );
 }
 
 export default BlankPage;
